DialCraft/requirements.txt
faster-whisper
sounddevice
scipy
numpy
ollama
python-dotenv
fastapi
uvicorn
websockets
pyngrok
webrtcvad
twilio
kokoro
